Split out a BaseListingView and template from IndexView
