+++ wagtail/admin/views/generic/base.py
@@ -2,10 +2,13 @@ from django.contrib.admin.utils import quote, unquote
 from django.core.exceptions import ImproperlyConfigured
 from django.shortcuts import get_object_or_404, redirect
 from django.urls import reverse
+from django.utils.translation import gettext_lazy as _
 from django.views import View
 from django.views.generic.base import ContextMixin, TemplateResponseMixin
+from django.views.generic.list import BaseListView
 
 from wagtail.admin import messages
+from wagtail.admin.ui.tables import Column, Table
 from wagtail.admin.utils import get_valid_next_url_from_request
 
 
@@ -113,3 +116,59 @@ class BaseOperationView(BaseObjectMixin, View):
         self.perform_operation()
         self.add_success_message()
         return redirect(self.get_success_url())
+
+
+class BaseListingView(WagtailAdminTemplateMixin, BaseListView):
+    template_name = "wagtailadmin/generic/listing.html"
+    results_template_name = "wagtailadmin/generic/listing_results.html"
+    results_only = False  # If true, just render the results as an HTML fragment
+    table_class = Table
+    columns = [Column("__str__", label=_("Title"))]
+    index_url_name = None
+    page_kwarg = "p"
+    default_ordering = None
+
+    def get_template_names(self):
+        if self.results_only:
+            return [self.results_template_name]
+        else:
+            return super().get_template_names()
+
+    def get_valid_orderings(self):
+        orderings = []
+        for col in self.columns:
+            if col.sort_key:
+                orderings.append(col.sort_key)
+                orderings.append("-%s" % col.sort_key)
+        return orderings
+
+    def get_ordering(self):
+        ordering = self.request.GET.get("ordering", self.default_ordering)
+        if ordering not in self.get_valid_orderings():
+            ordering = self.default_ordering
+        return ordering
+
+    def get_table(self, object_list, **kwargs):
+        return self.table_class(
+            self.columns,
+            object_list,
+            ordering=self.get_ordering(),
+            **kwargs,
+        )
+
+    def get_index_url(self):
+        if self.index_url_name:
+            return reverse(self.index_url_name)
+
+    def get_context_data(self, *args, **kwargs):
+        context = super().get_context_data(*args, **kwargs)
+
+        index_url = self.get_index_url()
+        context["index_url"] = index_url
+
+        table = self.get_table(context["object_list"], base_url=index_url)
+        context["table"] = table
+        context["media"] = table.media
+        context["is_paginated"] = bool(self.paginate_by)
+
+        return context
