+++ wagtail/snippets/templates/wagtailsnippets/snippets/index.html
@@ -19,7 +19,7 @@
     {% include 'wagtailadmin/shared/header.html' with breadcrumb=breadcrumb title=model_opts.verbose_name_plural|capfirst icon=header_icon search_url=search_url base_actions=base_action_locale action_url=action_url_add_snippet action_icon="plus" action_text=action_text_snippet %}
 
     <div class="nice-padding{% if filters %} filterable{% endif %}">
-        <div id="index-results" class="snippets">
+        <div id="listing-results" class="snippets">
             {% include "wagtailsnippets/snippets/index_results.html" %}
         </div>
         {% if filters %}
