+++ wagtail/users/templates/wagtailusers/groups/results.html
@@ -1,4 +1,4 @@
-{% extends "wagtailadmin/generic/index_results.html" %}
+{% extends "wagtailadmin/generic/listing_results.html" %}
 {% load i18n %}
 
 {% block no_results_message %}
